RDM8137 fix some encoding error on windows

+++ modules/EncryptionClient.py
@@ -53,6 +53,7 @@ class EncryptionClient(object):
         bag_dir_parent = None
         bag_output_path = None
         key_path_on_vault = None
+        sums_fullpath = None
         
         # generate key
         self._dataset_key_manager.generate_key()
@@ -67,8 +68,8 @@ class EncryptionClient(object):
         logger.debug("The temp directory created to hold the bag {}".format(bag_dir_parent))
         queue.put(bag_dir_parent)
 
-        try:
-            if os.path.isdir(self._input):
+        if os.path.isdir(self._input):
+            try:
                 # Generate checksums of each file in the directory
                 sums_fullpath = Util.generate_checksums(self._input)
                 
@@ -84,15 +85,13 @@ class EncryptionClient(object):
                 os.remove(zip_filepath)
                 # Remove the generated checksum file from the original data directory
                 os.remove(sums_fullpath)
-            else:
-                logger.error("Please select a directory to encrypt.")
-                return False
-        except Exception as e:
-            logger.error("Error in compressing and encrypting input directory")
-            if os.path.exists(sums_fullpath):
-                os.remove(sums_fullpath)
-            shutil.rmtree(bag_dir_parent)            
-            return False
+            except Exception as e:
+                if sums_fullpath is not None and os.path.exists(sums_fullpath):
+                    os.remove(sums_fullpath)
+                shutil.rmtree(bag_dir_parent)
+                raise Exception("Error in compressing and encrypting input directory")
+        else:
+            raise Exception("Please select a directory to encrypt.")
 
         try:
             bag = bagit.make_bag(bag_dir, None, 1, ['sha256'])
@@ -100,13 +99,12 @@ class EncryptionClient(object):
             bag.info['Dataset-UUID'] = dataset_uuid
             bag.save()
         except (bagit.BagError, Exception) as e:
-            logger.error("Error creating a bag of the encrypted package.")
             shutil.rmtree(bag_dir_parent)
-            return False
+            raise Exception("Error creating a bag of the encrypted package.")
         
         # create a file tree file
         try:
-            with open(os.path.join(bag_dir, "file_tree.txt"), "w") as f:
+            with open(os.path.join(bag_dir, "file_tree.txt"), "w", encoding="utf-8") as f:
                 tree = Tree()
                 tree.totalSize += os.path.getsize(self._input)
                 print(os.path.basename(os.path.normpath(self._input)), file=f)
@@ -114,9 +112,8 @@ class EncryptionClient(object):
                 print("\n" + tree.summary(), file=f)
                 bag.save()
         except Exception as e:
-            logger.error("Error in creating a file tree file: {}".format(e))
             shutil.rmtree(bag_dir_parent)
-            return False
+            raise Exception("Error in creating a file tree file: {}".format(e))
 
         # zip bag dir and move it to the output path
         bag_destination = os.path.join(
